requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
fastapi==0.115.14
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.10.0
markdown-it-py==3.0.0
mcp==1.9.1
mdurl==0.1.2
multidict==6.5.0
narwhals==1.41.0
numpy==2.2.6
openai==1.82.0
packaging==24.2
pandas==2.2.3
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
plotly==6.1.2
propcache==0.3.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.4
rich==14.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.46.2
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv==0.7.8
uvicorn==0.34.2
yarl==1.20.1
